# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: APEX INVESTOR API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /watchlist:
    get:
      operationId: getWatchlist
      tags: [watchlist]
      summary: Get watchlist
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: sector
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Watchlist items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WatchlistItem"
    post:
      operationId: createWatchlistItem
      tags: [watchlist]
      summary: Add to watchlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWatchlistItem"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchlistItem"
  /watchlist/{id}:
    put:
      operationId: updateWatchlistItem
      tags: [watchlist]
      summary: Update watchlist item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWatchlistItem"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchlistItem"
    delete:
      operationId: deleteWatchlistItem
      tags: [watchlist]
      summary: Delete watchlist item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    WatchlistItem:
      type: object
      properties:
        id:
          type: integer
        ticker:
          type: string
        companyName:
          type: string
        sector:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        addedDate:
          type: string
          nullable: true
        reason:
          type: string
          nullable: true
        targetPrice:
          type: number
          nullable: true
        currentPrice:
          type: number
          nullable: true
        status:
          type: string
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - ticker
        - companyName
        - status
        - createdAt
        - updatedAt
    CreateWatchlistItem:
      type: object
      properties:
        ticker:
          type: string
        companyName:
          type: string
        sector:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        reason:
          type: string
          nullable: true
        targetPrice:
          type: number
          nullable: true
        currentPrice:
          type: number
          nullable: true
        status:
          type: string
        notes:
          type: string
          nullable: true
      required: [ticker, companyName, status]
    UpdateWatchlistItem:
      type: object
      properties:
        ticker:
          type: string
        companyName:
          type: string
        sector:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        reason:
          type: string
          nullable: true
        targetPrice:
          type: number
          nullable: true
        currentPrice:
          type: number
          nullable: true
        status:
          type: string
        notes:
          type: string
          nullable: true