# lib/api-zod/src/generated/types/createWatchlistItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * APEX INVESTOR API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateWatchlistItem {
  ticker: string;
  companyName: string;
  sector?: string | null;
  country?: string | null;
  reason?: string | null;
  targetPrice?: number | null;
  currentPrice?: number | null;
  status: string;
  notes?: string | null;
}